requirements.txt
asgiref==3.11.1
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==26.0
PyJWT==2.12.1
python-dotenv==1.0.0
pytz==2026.1.post1
PyYAML==6.0.1
setuptools==82.0.1
sqlparse==0.5.5
uritemplate==4.2.0
wheel==0.46.3
